move action to near-vm-errors
