+++ core/primitives/src/lib.rs
@@ -6,11 +6,11 @@ pub use near_primitives_core::hash;
 pub use near_primitives_core::num_rational;
 pub use near_primitives_core::profile;
 pub use near_primitives_core::serialize;
+pub use near_vm_errors::{delegate_action, signable_message};
 
 pub mod block;
 pub mod block_header;
 pub mod challenge;
-pub mod delegate_action;
 pub mod epoch_manager;
 pub mod errors;
 pub mod merkle;
@@ -21,7 +21,6 @@ pub mod runtime;
 pub mod sandbox;
 pub mod shard_layout;
 pub mod sharding;
-pub mod signable_message;
 pub mod state;
 pub mod state_part;
 pub mod state_record;
