+++ runtime/near-vm-errors/Cargo.toml
@@ -13,10 +13,13 @@ publish = true
 [dependencies]
 borsh.workspace = true
 serde.workspace = true
+serde_with.workspace = true
 strum.workspace = true
 thiserror.workspace = true
 
 near-account-id.workspace = true
+near-crypto.workspace = true
+near-fmt.workspace = true
 near-primitives-core.workspace = true
 near-rpc-error-macro.workspace = true
