+++ runtime/near-vm-errors/src/action.rs
@@ -1,0 +1,238 @@
+use borsh::{BorshDeserialize, BorshSerialize};
+use near_crypto::PublicKey;
+use near_fmt::AbbrBytes;
+use near_primitives_core::{
+    account::AccessKey,
+    serialize::dec_format,
+    types::{AccountId, Balance, Gas},
+};
+use serde_with::base64::Base64;
+use serde_with::serde_as;
+use std::fmt;
+
+#[derive(
+    BorshSerialize,
+    BorshDeserialize,
+    PartialEq,
+    Eq,
+    Clone,
+    Debug,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct AddKeyAction {
+    /// A public key which will be associated with an access_key
+    pub public_key: PublicKey,
+    /// An access key with the permission
+    pub access_key: AccessKey,
+}
+
+/// Create account action
+#[derive(
+    BorshSerialize,
+    BorshDeserialize,
+    PartialEq,
+    Eq,
+    Clone,
+    Debug,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct CreateAccountAction {}
+
+#[derive(
+    BorshSerialize,
+    BorshDeserialize,
+    PartialEq,
+    Eq,
+    Clone,
+    Debug,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct DeleteAccountAction {
+    pub beneficiary_id: AccountId,
+}
+
+#[derive(
+    BorshSerialize,
+    BorshDeserialize,
+    PartialEq,
+    Eq,
+    Clone,
+    Debug,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct DeleteKeyAction {
+    /// A public key associated with the access_key to be deleted.
+    pub public_key: PublicKey,
+}
+
+/// Deploy contract action
+#[serde_as]
+#[derive(
+    BorshSerialize, BorshDeserialize, serde::Serialize, serde::Deserialize, PartialEq, Eq, Clone,
+)]
+pub struct DeployContractAction {
+    /// WebAssembly binary
+    #[serde_as(as = "Base64")]
+    pub code: Vec<u8>,
+}
+
+impl fmt::Debug for DeployContractAction {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("DeployContractAction")
+            .field("code", &format_args!("{}", AbbrBytes(&self.code)))
+            .finish()
+    }
+}
+
+#[serde_as]
+#[derive(
+    BorshSerialize, BorshDeserialize, serde::Serialize, serde::Deserialize, PartialEq, Eq, Clone,
+)]
+pub struct FunctionCallAction {
+    pub method_name: String,
+    #[serde_as(as = "Base64")]
+    pub args: Vec<u8>,
+    pub gas: Gas,
+    #[serde(with = "dec_format")]
+    pub deposit: Balance,
+}
+
+impl fmt::Debug for FunctionCallAction {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("FunctionCallAction")
+            .field("method_name", &format_args!("{}", &self.method_name))
+            .field("args", &format_args!("{}", AbbrBytes(&self.args)))
+            .field("gas", &format_args!("{}", &self.gas))
+            .field("deposit", &format_args!("{}", &self.deposit))
+            .finish()
+    }
+}
+
+/// An action which stakes signer_id tokens and setup's validator public key
+#[derive(
+    BorshSerialize,
+    BorshDeserialize,
+    PartialEq,
+    Eq,
+    Clone,
+    Debug,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct StakeAction {
+    /// Amount of tokens to stake.
+    #[serde(with = "dec_format")]
+    pub stake: Balance,
+    /// Validator key which will be used to sign transactions on behalf of signer_id
+    pub public_key: PublicKey,
+}
+
+#[derive(
+    BorshSerialize,
+    BorshDeserialize,
+    PartialEq,
+    Eq,
+    Clone,
+    Debug,
+    serde::Serialize,
+    serde::Deserialize,
+)]
+pub struct TransferAction {
+    #[serde(with = "dec_format")]
+    pub deposit: Balance,
+}
+
+#[derive(
+    BorshSerialize,
+    BorshDeserialize,
+    PartialEq,
+    Eq,
+    Debug,
+    Clone,
+    serde::Serialize,
+    serde::Deserialize,
+    strum::AsRefStr,
+)]
+pub enum Action {
+    /// Create an (sub)account using a transaction `receiver_id` as an ID for
+    /// a new account ID must pass validation rules described here
+    /// <http://nomicon.io/Primitives/Account.html>.
+    CreateAccount(CreateAccountAction),
+    /// Sets a Wasm code to a receiver_id
+    DeployContract(DeployContractAction),
+    FunctionCall(FunctionCallAction),
+    Transfer(TransferAction),
+    Stake(StakeAction),
+    AddKey(AddKeyAction),
+    DeleteKey(DeleteKeyAction),
+    DeleteAccount(DeleteAccountAction),
+    Delegate(crate::delegate_action::SignedDelegateAction),
+}
+
+impl Action {
+    pub fn get_prepaid_gas(&self) -> Gas {
+        match self {
+            Action::FunctionCall(a) => a.gas,
+            _ => 0,
+        }
+    }
+    pub fn get_deposit_balance(&self) -> Balance {
+        match self {
+            Action::FunctionCall(a) => a.deposit,
+            Action::Transfer(a) => a.deposit,
+            _ => 0,
+        }
+    }
+}
+
+impl From<CreateAccountAction> for Action {
+    fn from(create_account_action: CreateAccountAction) -> Self {
+        Self::CreateAccount(create_account_action)
+    }
+}
+
+impl From<DeployContractAction> for Action {
+    fn from(deploy_contract_action: DeployContractAction) -> Self {
+        Self::DeployContract(deploy_contract_action)
+    }
+}
+
+impl From<FunctionCallAction> for Action {
+    fn from(function_call_action: FunctionCallAction) -> Self {
+        Self::FunctionCall(function_call_action)
+    }
+}
+
+impl From<TransferAction> for Action {
+    fn from(transfer_action: TransferAction) -> Self {
+        Self::Transfer(transfer_action)
+    }
+}
+
+impl From<StakeAction> for Action {
+    fn from(stake_action: StakeAction) -> Self {
+        Self::Stake(stake_action)
+    }
+}
+
+impl From<AddKeyAction> for Action {
+    fn from(add_key_action: AddKeyAction) -> Self {
+        Self::AddKey(add_key_action)
+    }
+}
+
+impl From<DeleteKeyAction> for Action {
+    fn from(delete_key_action: DeleteKeyAction) -> Self {
+        Self::DeleteKey(delete_key_action)
+    }
+}
+
+impl From<DeleteAccountAction> for Action {
+    fn from(delete_account_action: DeleteAccountAction) -> Self {
+        Self::DeleteAccount(delete_account_action)
+    }
+}
